small fixes, bring tests to passing
add-room-to-group can't be a target-step since the state event doesn't
exist for the group yet, so it has been changed to a room-step.

regex updated for add-to-group parser so that whitespace doesn't mess
thing up because cl-strings:split "foo baz " -> "foo" "baz" "" which meant
we'd pass a target room with an id of "" to add-room-to-group. A test
has been added for this stuff too.

fixed an issue where I was redefining execute for job instead of
value-job in value-step.lisp.