clear up command-parsers
changed &rest lambda-list in command parser to fixed 4 symbols to enforce
the pattern that we were using (and should be expected by the
luna-command hook.

added a default report to steps defined with define-command-parser
to print the name of the command to the stream (and not use the
default single-step reporter which would print the args and thus the
entire event which would create a lot of noise).