ban and redact all
these both work but redact all has some serious issues if you're rate
limited. which suggests my approach to this entire thing is wrong.

if you rate limit to like 0.2 (default), then cpu gets to 100%, I've
checked the header stream, I've run the profiler. The issue seems to
be in the way we're using lparallel.