make steps to make better use of async.
Return (cons room-id error) from steps. This change will ofc have
introduced ugly looking reports, but reports are to change soon and
make use of the room preview.

There's probably a whole bunch of luna-lambda being used when we're
probably not going to want to anymore.

channel has been moved out of the listener so that it can be GCd as
the resutls queue was never being popped.

the plan next is to bring back blackbird and use this for promises
instead of lparallels promises since they aren't as easy to work with
imo.